requirements.txt
AWSIoTPythonSDK==1.5.4
bcrypt==4.1.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
CoAPthon3==1.0.2
colorzero==2.0
cryptography==42.0.5
grpcio==1.62.1
grpcio-tools==1.62.1
idna==3.7
importlib==1.0.4
kafka-python==2.0.2
paho-mqtt==2.0.0
paramiko==3.4.0
pika==1.3.2
protobuf==4.25.3
psutil==5.9.8
pycparser==2.22
PyNaCl==1.5.0
pyzmq==25.1.2
requests==2.31.0
urllib3==2.2.1
websocket-client==1.7.0
zmq==0.0.0
